requirements.txt
annotated-types==0.5.0
anyio==3.7.1
async-timeout==4.0.2
click==8.1.6
fastapi==0.101.0
greenlet==2.0.2
h11==0.14.0
httptools==0.6.0
pydantic==2.1.1
pydantic-settings==2.0.2
pydantic_core==2.4.0
pylint==2.17.5
PyMySQL==1.1.0
python-dotenv==1.0.0
PyYAML==6.0.1
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.19
starlette==0.27.0
typing_extensions==4.7.1
uvicorn==0.23.2
uvloop==0.17.0
websockets==11.0.3
google-generativeai==0.5.0